+++ util/noise/spot_noise_gd.h
@@ -1,12 +1,14 @@
 #ifndef ZN_SPOT_NOISE_GD_H
 #define ZN_SPOT_NOISE_GD_H
 
+#include "../godot/classes/noise.h"
 #include "../godot/classes/resource.h"
+#include "spot_noise.h"
 
 namespace zylann {
 
-class ZN_SpotNoise : public Resource {
-	GDCLASS(ZN_SpotNoise, Resource);
+class ZN_SpotNoise : public ZN_Noise {
+	GDCLASS(ZN_SpotNoise, ZN_Noise);
 
 public:
 	int get_seed() const;
@@ -21,14 +23,21 @@ public:
 	float get_jitter() const;
 	void set_jitter(float jitter);
 
-	float get_noise_2d(real_t x, real_t y) const;
-	float get_noise_3d(real_t x, real_t y, real_t z) const;
+	inline float get_noise_2d_inline(const Vector2f pos) const {
+		return SpotNoise::spot_noise_2d(pos, _cell_size, _spot_radius, _jitter, _seed);
+	}
 
-	float get_noise_2dv(Vector2 pos) const;
-	float get_noise_3dv(Vector3 pos) const;
+	inline float get_noise_3d_inline(const Vector3f pos) const {
+		return SpotNoise::spot_noise_3d(pos, _cell_size, _spot_radius, _jitter, _seed);
+	}
 
 	PackedVector2Array get_spot_positions_in_area_2d(Rect2 rect) const;
 	PackedVector3Array get_spot_positions_in_area_3d(AABB aabb) const;
+
+protected:
+	real_t _zn_get_noise_1d(real_t p_x) const override;
+	real_t _zn_get_noise_2d(Vector2 p_v) const override;
+	real_t _zn_get_noise_3d(Vector3 p_v) const override;
 
 private:
 	static void _bind_methods();